Validar registro local
Quedan los editar
